Alter the TikZ base function to stop using the size member of nuclide.
The default scale is cm so even with a size=3 for A5, the full chart
would be 5.3x3.5m. This was causing the "Dimension too large"
compilation error with latex. Changing to a scale of mm caused issues
with the fonts and other placements. Until I get more knowledge of
TikZ we will always use the same scale (1) and the user can scale the
picture in their document.