Do not need to calculate drip line values for isomeric states.
As the file is ordered in A, looping backward through already
recorded isotopes, means once we have calculated all necessary values or
the difference in A is >3 we can stop checking as we know that further
data will not be used. Tests show that this will reduce the number of
times we loop through the vector of isotopes by ~98%.

Also, STL containers do not need to be initialised to empty. This is
done by the default constructor.