Move Z->Symbol, Symbol->Z and setExtreme() functions to members.
Have also got rid of the symbol member of the Nuclide class as I
could not see a way to make it work once the conversions moved to
member functions, and this was the only time it was used. The
conversion to symbol is now called on demand as we write the
output file.

The conversion functions are bascially big maps so have been put
in their own files to avoid polluting the other functions of the
inputs class. They should probably be members of Nuclide or free
floating for global use, however they are currently only used
when dealing with the inputs class which is why I have put them in
there.