Move reading of R-process data to inputs member function
The R-process data was tidied into the writing of the EPS file, and
the file was read twice. Reading is now in it's own member function
so is independent of file type. It is also stored in a vector of
pairs which can be accessed the second time the data is needed, in
the case oe an EPS file, rather than re-reading the file.