Use the MAX_Z const rather than a magic number and count lines after the file is opened.
Also as most of the file is in an if/else loop, change the pattern from

if(true)do something
else exit

to
if(!true)exit
do something

This should remove a lot of indenting. Not really important, but there
are some quite deep 'nests' so am looking to make things as shallow
as possible.