Reading of NUBASE_ME was wrong, the numbers 'copied' when moved to a member were incorrect.
Alter how drip values are calculated; point to back of Nuclide vector for comparison rather
than using an iterator.

Use sscanf to read the r-process values as the line format is simple

Add and change some valgrind options, present in the Makefile as comments and used for
checking/testing/validation.