Stop using Nuclide objects when we only use a fraction (~0.25) of the values.
Currently used version of GCC/C++ against which we compile (c++03) does not
allow vectors of local variables so had to move the struct 'outside' the
function. The file which is read will change so cannot use an array and I
don't want to just create a massive array that  will always be large enough.