Initialise isomer energy value and make sure lowest partition is positive.
There is no need to set the gs isomeric data to an 'obvious' value, lets
initialise it to <0 even though we can filter out gs by checking value
of the st member.

As a follow up to new initialisation, we need to ensure that only
positive values are displayed in the colour of the first element of
partition object.