Make nesting shallower by checking for unwanted condition and exiting if true.
It seems that emacs uses a combination of <space> and <tab> when
you press the TAB key, lets starting converting to all spaces.
Currently I'll make the change for each file in which an alteration
is made, but I may just go through and change all files if I have
the time one day.