Pass 2 doubles instead of the entire vector of Nuclides.
When calculating the drip-lines, using the finite range drop model,
we were passing the full vector of Nuclide classes (~4.5k elements)
many times, just to have access to the mass excess values for a
neutron and proton. Now we pass on the two values that we need.

Untabify more of the files.