Add explanatory comments to the eps file
If the key is taller than the range of Z to be drawn, we shift the chart
position. The brevity of the eps syntax meant that I thought I had left
a dangling restore state. I hadn't, but it took some investigation so
add comments to avoid this in the future.