Don't print warning message if -f option is after -o
We check to see if the extension has been given on the optional output
file name (-o) and as part of this set the file type (-f). As a result
we then don't fall into the outer most check for -f so print the unknown
option warning message. This addition does not print the message for the
situation described.

If we read -o before -f, there is no need to check all console options
again, we can start from out current position.