The table year can now be set via console option
This opened up a real can of worms. There was an additional value that
was introduced after 2003 so a reordering of the inital part of the
code was required. Summary of the over-arching changes is below:

- The mass table was read before any console options were processed, thus
a large re-write was needed. In doing so, some free functions were moved
to be methods of the input class and other related functions were tidied
up, or removed.

- The file type has been extracted to an enum, rather than an int member
of the input class.

- createChart is no longer overloaded, we store the validity of the
input arguments so can have a single use for createChart.

- The mass table year is now used to extract the decay string. We
started to implement this in be408719a, but didn't do it properly.

- C++11 provides stoi(), so use that instead of atoi() as we don't use
const char*.