Standardise behaviour between scripts
- Python made a single chart if the number wasn't specified but BASH
printed the help and exited. Both will now create one chart if no number
is given.

- The python script was looking in the wrong place for the default
executable created by cmake. Now fixed.

Generalise search for the executable

We now look for the executables relative to the scripts directory,
rather than PWD. This should allow it to be found in the majority of
scenarios.

ascii_lowercase has 1=a not 0=a, make necessary alterations

Use the ternary operator