Move key functionality to it's own class
For better encapsulation we now don't know the size of the chart at
the time of writing the prolog. Luckily, the people who wrote
postscript envisioned this scenario so we can just specify that the
size will be given at the end of the file.

Was also able to start tiding up the cluttered inputs class.