Add .clang-format file
Indenting and brace style can be a hot topic, take away any debate
by using clang-format.

Only time we switch off clang-fromat is for if() conditions that
contain multiple conditions. This is what I would like:

if (AAA
    && BBB
    && CCC)

but there is no setting for this so it's converted to

if (AAA && BBB && CCC)

or, gets broken at <ColumnLimit> if line is too long.