Only push codecov data from 1 CI, now that I think I've figured it out
This is a subtle one, but I think each 'run' line in circleCI launches
it's own shell so we were never in the correct directory while running
lcov.

For travis, that was my mistake, Release build had just been run so we
were still in that directory. Coverage data is for the Debug build.