Remove TODO item, not needed
The add_executable() call uses cmake variables so we would need to
look for the main project() call. I've split this onto multiple
lines so would need to look for the lines after. That would be too
fragile so lets just stick to what we are doing for the time being.