Major overhaul of the dripline class and the files it reads/creates
- Disentangle the reading of the drop model file from the creation of
the drip line files that are read by the poject. The drop model file
now has it's data stored in a static member so will only be read once,
no matter which drip lines are drawn and which files do no exist.
- A generic function has been created that will return the N/Z value
when the required separation energy has gone negative, creation of the
drip line files leverage this function.
- Separation energy at the drip line is no longer recorded. It was
never used so is not needed. The necessary test has been updated as
required.

Fix bug in the reading of the option file that was showing a warning
when it should have been.