Start trimming down the Option class
It currently contains data that while technically an option, cannot be
changed once the program is run. Start moving some of those things to
the class they belong to. This should also mean we do not need to pass
an instance of the Option class to almost every other class used to
draw the chart.