Refactor getting non-file type specific values
For the case of colouring by first isomer value, we drawn an empty
isotope for those with no data as a visual aid. These are marked with
a flag and were drawn a special way. This does not seem to be
necessary and be combined with 'normal' drawning.

I have tested the above and am almost certain it's true. Hopefully
past me just missed something and I am not introducing a bug.