Use an object library to avoid duplicate compilation
Listing the implementation source files as dependencies of the test
binary, although correct, meant that they were being compiled
twice. Once for the main project binary and once for the test
binary. When test coverage is complete this would mean 50% extra work
and compilation time.

The object library in cmake allows the implementation files to be
compiled into a single library and thus reused when created the test binary.